Time: 1 ms (30.25%), Space: 7.8 MB (61.42%) - LeetHub

+++ 2649-count-total-number-of-colored-cells/2649-count-total-number-of-colored-cells.cpp
@@ -2,11 +2,11 @@ class Solution {
 public:
     long long coloredCells(int n) {
         long long count = 0;
-        for (int i = n - 1; i >= 0; i--) {
+        for (long long i = n - 1; i >= 0; i--) {
             if(i == n - 1){
-                count += 2 *i + 1;
+                count += 2 * i + 1;
             }else{
-                count += 4 *i + 2;
+                count += 4 * i + 2;
             }
         }
         return count;
